Re-organise some of the code
